pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapiproj-orders-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "faststream[rabbit]>=0.6.6",
    "greenlet>=3.3.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis[async]>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["Orders*", "Users*", "core*", "database*", "jwt_auth*", "cert*"]
